refactor: Repackage exceptions
예외 클래스들 패키지 세분화
