--- src/main/java/kr/co/khacademy/semi/exception/AccountIdNotFoundException.java
@@ -1,4 +1,0 @@
-package kr.co.khacademy.semi.exception;
-
-public class AccountIdNotFoundException extends LoginFailedException {
-}
