--- src/main/java/kr/co/khacademy/semi/exception/LoginFailedException.java
@@ -1,4 +1,0 @@
-package kr.co.khacademy.semi.exception;
-
-public class LoginFailedException extends RuntimeException {
-}
