--- src/main/java/kr/co/khacademy/semi/exception/PermissionIdNotFoundException.java
@@ -1,4 +1,0 @@
-package kr.co.khacademy.semi.exception;
-
-public class PermissionIdNotFoundException extends LoginFailedException {
-}
