--- src/main/java/kr/co/khacademy/semi/exception/RoleIdNotFoundException.java
@@ -1,4 +1,0 @@
-package kr.co.khacademy.semi.exception;
-
-public class RoleIdNotFoundException extends LoginFailedException {
-}
