--- src/main/java/kr/co/khacademy/semi/exception/UsernameNotFoundException.java
@@ -1,4 +1,0 @@
-package kr.co.khacademy.semi.exception;
-
-public class UsernameNotFoundException extends LoginFailedException {
-}
