+++ src/main/java/kr/co/khacademy/semi/exception/login/LoginFailedException.java
@@ -1,0 +1,11 @@
+package kr.co.khacademy.semi.exception.login;
+
+public class LoginFailedException extends RuntimeException {
+    public LoginFailedException() {
+        super();
+    }
+
+    public LoginFailedException(String message) {
+        super(message);
+    }
+}
