+++ src/main/java/kr/co/khacademy/semi/exception/login/sub/AccountIdNotFoundException.java
@@ -1,0 +1,6 @@
+package kr.co.khacademy.semi.exception.login.sub;
+
+import kr.co.khacademy.semi.exception.login.LoginFailedException;
+
+public class AccountIdNotFoundException extends LoginFailedException {
+}
