+++ src/main/java/kr/co/khacademy/semi/repository/PermissionRepository.java
@@ -1,7 +1,7 @@
 package kr.co.khacademy.semi.repository;
 
 import kr.co.khacademy.semi.entity.Permission;
-import kr.co.khacademy.semi.exception.PermissionIdNotFoundException;
+import kr.co.khacademy.semi.exception.login.sub.PermissionIdNotFoundException;
 
 public class PermissionRepository {
